backend/requirements.txt
# Core
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL driver

# ML
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Auth
pyjwt>=2.8.0
requests>=2.31.0  # Google token verification

# Utils
python-dotenv>=1.0.0
Pillow>=10.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
